Allowing multiple quantities of variants to be ordered in one step
Orders controller expects array of form << quantities[product] = "var_id=nnn" >>

This allows nnn items of a single variant of a product to be ordered, and also
for several products to be ordered at the same time.

Note: this coding is needed to ensure correct radiobutton functionality
(choosing only one variant), and is insufficient to represent simultaneous
orders of several variants of the same product.