Fixed spree/spree#465: made guest checkout the default option
This give the appearance that Spree is remembering the registration choice. If the user has selected either
of the other options, an Ajax call will log them in and this section won't be shown when they next reload
the page. Hence if the user has not selected these (and not logged in) we can assume they do want to
proceed as a guest.