660 - add default addresses for registered users
This is the simplest form of address book: each user contains their last
used ship and bill addresses, and these are used as defaults in their
next checkout. If any edits are made, these are saved as new records -
the previous addresses are not modified (though administrators can
still fix errors if needed). At the end of checkout, the (ids of the)
resulting addresses are saved to the user.

Extracted from swistak's work on spree/spree#54, with a few simplifications,
in particular that the checkout controller determines whether or not
a fresh address is needed - since it has all of the relevant info,
and that addresses can only be edited in the checkout page.

It is expected that further address book functionality will be added
as an extension.

[#600 state:resolved]