661 - get checkout forms to show state_name
it will start with a text field for state name when the address country
has no states, else start with a select box, and ensure that the widget
has the correct initial state value. (Remember that update_state in
checkout.js modifies these widgets...)

this change also incorporates a preference :address_requires_state,
which when true will remove the state fields from the checkout address
forms and allow the state part of an address object to be blank.

[#661 state:resolved]