696 - product clone function for admin
This allows quick entry of similar products by cloning an existing product.

All of the source product's attributes are duplicated, and the taxons,
properties, and image structure is cloned. The product's master variant
is always cloned and its images cloned. If the product has (real) variants,
just the option types are copied.

Products may have application-specific fields, so there's a hook
"duplicate_extra" which can do extra adjustments on the cloned
object before it gets saved.

The permalink will need to be adjusted via other means.

[#696 state:resolved]