Speed up ProductGroup test
Previously this test case was inheriting from vanilla Test::Unit::TestCase,
which does not wrap each test in a transaction.  Because of this there was
logic to delete the objects and use of find_or_create_by.  By using
ActiveSupport::TestCase we get each test wrapped in a transaction and it is
no longer necessary to have the extra logic.

[#1044 state:resolved]