Fix to make Country and State controls work in IE8, IE7 and IE6
This commit fixes an issue in the Address page of the checkout process.  The Country and State controls were not working as expected in IE8, IE7, or IE6.  The State control was displaying as a textbox when United States was selected as a country.
The fix consists of making the jQuery search expressions more specific.  That way jQuery can find the controls it needs to manipulate when the user selects a country.
For some reason, the existing code was working in other major browsers (Chrome, Safari, Firefox).  This fix preserves the behavior in those browsers.