Move code dealing with customer details out of Admin::OrdersController
This has how been moved into Admin::Orders::CustomerDetails to lessen the clutter in the OrdersController. imo, the OrdersController should only deal with *orders*, and while customer details *are* related and are dealt with by calling update_attributes on an Order instance, I think it deserves its own controller because it's different information.