Run 'bundle install' rather than 'bundle update'
The reasoning behind this is because 'bundle update' will delete the Gemfile and re-'work out' the dependencies for the application. This would include Bad Things(tm) like it upgrading the version of a gem you had locked at 2.6 to 2.7, for instance. That *could* introduce breaking changes, and so it's something that we should try to avoid.

What 'bundle install' will do in this case is install the gems that it *doesn't* have, leaving the ones it *does* have well and truly alone.