Add store/helpers.js.erb
This file provides a Spree JS container for any settings that we need to make available to Spree's JavaScript components. Right now, that's just base_path.

The base_path helper provides the location as to where Spree is mounted. This is needed for any URLs that may wish to be hard-coded, like the ones used in spree/spree#1029.

This commit will fix spree/spree#1029