Fix broken image sorting in admin section. Closes spree/spree#1101.
This was broken due to the namespace changes not being applied to the ids for this form.

For this change, I've 'hard-coded' the URLs as data attributes on the tables themselves. This means that the JS will not need to infer where the routes go from the id any more, which really was a hack. This also opens up the possibility of using nested routes, as is required in the images area for the backend admin.