Define precompile_assets! method inside Spree::Core::Engine
This method will need to be called if app.config.initialize_on_precompile is set to false inside the application. This is because with this option set to false *no* initializers inside application or engine will be run... including our little hook to add some assets to precompile.

This therefore causes the Spree assets that we so deeply care about to be blatantly ignored by the gorgeous Asset Pipeline. So we hack around it.

Underneath config.assets.initialize_on_precompile = false in your config/application.rb put this line:

Spree::Core::Engine.precompile_assets!

And all will be well.

P.S. YEEHAW!

Fixes spree/spree#1159