better imagemagick check suggested by Avdi
While pairing with Avdi Grimm he noticed the return value from %x() is
not used. Instead the variable $? is checked. A better approach is using
'system' which will return true if the command executed successfully and
then checking that value. It removes the need to check $? and also gets
rid of the need for rescue. Another benefit is that testing this part of
the code becomes easier.

Merges spree/spree#1660