Re implement customer searching in order backend
Fixes spree/spree#1702.

User models will need to respond to a Ransack call and have relationships to ship_address and bill_address models (which they do if they're using Spree), and we'd really really appreciate it if it had an email field on the table too.

Other than that, this feature works exactly the same way that it worked in 1-1-stable.