Prevent multiple checkout flow transitions
The issue was that the previous state_machine definition was not correctly being removed, causing the callbacks for states to remain inplace, when really they should be reset every time checkout_flow is defined. There is now a test within spec/models/order/checkout_spec.rb which re-defines the checkout_flow within the Order class and then validates that the next event "branches" don't know anything about the states that shouldn't be there. This should be sufficient for testing that old states are removed.

This commit also fixes a number of other, related problems:

1) Subclasses of the Order class should not effect the parent class's state_machine definition. This was fixed by using class_attribute instead of using cattr_accessor. Really, I (Ryan) should've known better.

2) checkout_steps was being incorrectly defined as an array, when it should be an ActiveSupport::OrderedHash instead

3) Prevented method conflict errors by using StateMachine::Machine.ignore_method_conflicts. This was only needed due to changes from this commit.