Move updating logic out of Order model and into its own object
This makes it easier to test, as well as providing direct access to the methods for updating shipment and payment state, for situations where they may be modified outside the context of an order update and may need to be updated independent of that. Situations such as when payments are processed.