Add currency support to models.
Fixes spree/spree#2197

Add currency parameter support to Spree::Money.

Spree::Money now allows an optional currency parameter to be passed in,
which will be used if specified.  In the case where it is not specified,
the global spree configuration value will be used instead.

The foundation of multi-currency support is going to be modifying the
code to support storing the currency in the various objects used by
Spree.

The goal is to modify the code to allow:

- adjustments
- line items
- orders
- payments
- return authorizations
- shipments
- shipping rates
- variants

to eventually store and track their own currency.

Currently these objects mirror the global currency value.  As code
support improves, this will be stored in the database for these models,
or inferred from dependent models (for example, determining the line
item currency from the order currency).

Flagged money helper as deprecated

The money helper is based around having a single site wide currency.
This is instead being moved to a series of display_* calls on the models
which have both the monetary amount and currency to produce its own
Spree::Money object, which then allows us to display a properly
formatted currency on the output.

Split prices from variants & store currency in DB.

This will allow a variant to have many prices in many different
currencies.

Store currency in database.

Add currency to line items.

Line items now support currency, which can be tracked on a line
item basis.

Add currency support to orders.

Infer payment currency from order.

Support currency on shipping.

Base currencies off of order currency.

Adjustment, Return Auth, and Shipment now infer their currency from the
currency of their order.