Show add to cart if price in selected currency.
This is a bug with the cart form.  It would show the add to cart button
if there was a valid price in the master currency rather than the
current currency.  This causes the add to cart button to not display if
the product is not available in the master currency but available in a
different currency.

Fixes spree/spree#2275