Allow order to be transitioned by calling next API endpoint
Previously, this was done through the update action, but now since that the order.next return value isn't checked there (for good reasons!) we're going to have to do it somewhere else. A 'next' endpoint is a logical choice, so we do that. This will also be able to tell users *why* their orders may not be transitioning in the API