Add translation to Product permalink
Also override the Globalize::ActiveRecord::Migration module in
globalize3 gem because the one in version 0.3.0 doesn't allow us to add
translations to a field in a existing table. No new version of
globalize3 has been released so far.

The module added here includes a `add_translations_fields!` method that
makes it possible to add a new migration file instead of overriding the
other one already checked in the repo.