Refactor frontend address spec
This spec has failing on the CI for a while. The reason might be that it
used capybara `find` to check the visibility of elements and that method
does not wait until the element become either visible or not visible so
depending on the page load time the spec may get confusing results.

`has_selector?` is probably a better choice since we can specify
explicitly the `visible` option

[Fixes spree/spree#3084]