Remove `return` from rescue_from block in Auth helper
Specs were failing due to a change on AS::Rescuable. See following rails
ref for details:

    https://github.com/rails/rails/commit/9c857db75788c21f6184279c130e79f21c750f9f
    https://github.com/rails/rails/commit/d1887d384a4038d67bcad50a46a6dc67abfebe04

This is the error you get, as of Rails 4, when calling `return` from
within a `rescue_from` block:

    LocalJumpError: unexpected return

According to ruby-doc the error raises when Ruby can't yield as
requested. e.g. you can't return inside a proc and then call it

    def get_me_a_return
      Proc.new { return 42 }
    end
    get_me_a_return.call