Fix backend / dash admin_path conflict
The named route `admin` has been renamed to `admin_overview` on the dash
gem and the the admin_path put back into backend gem. Because the
backend gem depends heavily on the `admin_path` helper and I don't think
it should depend on the dash gem I figured it just safer to rename the
named route on the dash gem instead.

See https://github.com/rails/rails/pull/9704 for details on why (?!) the
app now breaks when the same named route is defined twice