'Correctly' define zones association on 'create_shipping_method_zone' migration
Fixes spree/spree#3307

The issue happened upgrading a 1.3 app to 2.0. What would happen is that this migration expected the table to be called shipping_methods_zones, but a later migration (and a model change in the shipment model) changes this to being spree_shipping_method_zones. This small model code here will redefine the association and appease the migration running script.