Change UserLoggedInRule to use presence of a user as its check
The logic goes that if the user is signed in, then the order will be associated to that user (through the magic of associate_user). Therefore, if an order has a user then the user has been signed in at some point in the past. If the user is not signed in, only 'email' would be set for the user

Conflicts:
	core/app/models/spree/promotion/rules/user_logged_in.rb