Promo total is no longer factored into Order#shipment_total
This is because to calculate the order total, we take the item total (with no promos), the shipment total (with no promos) and then apply the total of the adjustments. Applying the promo_total within the shipment total calculation means that the order total has that adjustment applied effectively twice.