Ensure shipment costs are calculated before their promotions
This was an ordering bug, where the shipments promotions were being calculated before a shipment's final cost was set. This was causing Free Shipping promotions to be applied, but they would result in a $0 adjustment.

Switching the order in which these things happen makes it work.

I've also shored up the test for this within the free_shipping_promotions_spec.rb in the frontend.

Fixes spree/spree#4428