Fixes entering localized amounts for payments
While running Spree in Dutch locale, adding a Payment recognizes the decimal separator as thousands separator. So when using the pre filled value (e.g. 45.21) this will be converted to a much larger amount (4521.00). The fix makes sure that a localized value is filled in (45,21 instead of 45.21).

Tested on Spree 2.2.1

Fixes spree/spree#4753

Conflicts:
	backend/app/views/spree/admin/payments/_form.html.erb