Adding data-hooks to search results header
Our app modifies the look of the page when no results are found, and in order to do that we have to override this entire partial. I'm thinking that adding some data-hooks might make it easier for folks like us to modify just a part of the partial if needed.

Fixes spree/spree#4820