Make it possible to select a default stock location
When creating returns, the stock locations are currently shown in
whatever order the database returns. This allows designating a default
stock location, which will always show up first. That makes ninjas'
lives easier as they don't have to remember to choose the right location
every time.

For context: When they pick the wrong location that is currently the
default they just hit a 500.