Default credit card per user
When an order is completed, we save the credit card used as the user's
default credit card. When creating a new order, we use an existing
credit card (if it exists) and attach it to the order before reaching
the payment state.

cherry-picked from e891bf56dcbe188dd485e1a95da8445c23a9000e in spree master