Revert ReimbursementItem model
This removes ReimbursementItem model and squishes it back into the ReturnItem model.
We split this model out of ReturnItem and we still think that's a good idea.  But we've run into some issues with it so we'll have to look at adding it back in later when we can work through all the details.

Conflicts:
	core/spec/models/spree/reimbursement_spec.rb

cherry-picked from 5c6cbf9e49269e1a6aa6f472c5a7b2501fd321d6 in spree master