Add API endpoint /api/shipments/mine
With expedited exchanges we now have times when stores will be interested in deep details of a single shipment, without having to grab all the details of all shipments in an order.

Trying to prevent a bunch of n+1 queries with a bunch of includes.

Conflicts:
	api/config/routes.rb
	core/app/models/spree/shipment.rb

2-2-dev changes:

  - restrict "2-2-dev"-specific before filters

cherry-picked from pending PR in spree master: https://github.com/spree/spree/pull/5219