Fix spec
An in-memory timestamp was being compared to a db-truncated timestamp and
giving unexpected results.

It might be better to:

1. Find a more general solution to this so it doesn't bite us elsewhere.
2. Provide a more explicit way to record unreturned exchanges than comparing
timestamp fields values.

But I couldn't think of something great for 1, and 2 would expand the scope of
this a lot.

More details:  An in-memory order.created_at timestamp was being compared to a
db-truncated shipment.created_at timestamp.  Rails' sql adapter truncates
database times at the microsecond level because Time#usec truncates instead of
rounding:

https://github.com/rails/rails/blob/v4.0.13/activerecord/lib/active_record/connection_adapters/postgresql/quoting.rb#L163

e.g.:

    >> Time.at(0.19999999999).usec
    => 199999