Fix bug with mixed variant stocking during purchase.
Here is the bug:
1) An order is placed for two items, 1 of which tracks inventory, and
the other of which does not track inventory at the first variant's stock
location.
2) The shipment building process packs both items into a shipment for
the stock location for the first item, even though the second item has
no stock location, due to incorrect checking of configured stock in the
Stock::Coordinator.
3) Shipment#finalize! calls manifest_unstock, which then calls
StockLocation#move, which will go ahead and create a stock item for the
variant / stock location in order to do a stock movement.

This results in:
1) The variant that does not track inventory being fulfilled from a stock
location that it is not configured to be fulfilled from.
2) That same variant now has a stock item in the stock location it
previously was not configured to be fulfilled from, and can easily be
purchased from it going forward.

The fix put in place here is:
1) Fix the bad logic in the Stock::Coordinator to only pack inventory
units to the stock locations that the variant has a stock item for.
2) Disallow the creation of new stock items for variants in a stock location
when making a non-positive stock movement.

Conflicts:
	core/app/models/spree/stock/coordinator.rb
	core/app/models/spree/stock_location.rb
	core/spec/models/spree/stock_location_spec.rb