Allow receipt of stock transfer transfer items
Uses the variant autocomplete search component to select a variant. When
the search term matches exactly one variant, the receipt of that variant
against the current transfer order will automatically be triggered. When
multiple variants match the search term, the admin will have to select a
variant from the returned results. This ensures that items from stock
transfers will be processed individually and decreases the likelihood of
incorrectly processing inventory on receipt.

Additionally, the variants returned in the variant search component are
*not* filtered to only contain the variants that are part of the stock
transfer. If a variant is included in the stock transfer that is not
expected, it should be handled separately (e.g. adjust the quantity in
the stock item management section).

When a variant is received successfully, there is however a
way to update the received quantity if the stock transfer contains a large
quantity of the same variant or to correct any possible input errors.

Conflicts:
	api/app/helpers/spree/api/api_helpers.rb
	api/config/routes.rb
	backend/app/views/spree/admin/shared/_routes.html.erb
	backend/config/routes.rb