Replace order strings with symbols/arel_table.
There are some issues that some users can encounter to to a lack of
specific naming due to SQL errors on ordering. If the scope has
something like `.order('name')` it may cause a SQL error when we join
with another table that also has a name column.

As part of this, I got rid of a couple of generally terrible tests which
rely VERY heavily on their specific implementation.