Refactor gateway_spec for clarity and conciseness
the no_card payment method made it unclear what was being tested and
didn't add any value.

I'm not 100% sure that the third test testing both not completed orders
and cc's not associated to the user has value, but it seems like CC"s
don't necessarily need to be associated to a user.