Strip out ssl options from settings / codebase
ssl_required and ssl_allowed are historical relics of yore.

In 2014 any ecommerce site should enforce SSL on all connections. This
can either be set in rails, or handled at a web server/load balancer
layer, and the logic does not need to leak into the spree engine.

To handle in rails see:
http://api.rubyonrails.org/classes/ActionController/ForceSSL/ClassMethods.html

Conflicts:
	backend/app/controllers/spree/admin/general_settings_controller.rb