Drop version requirement on jquery-rails
rails installs with no explicit requirement on the version of jquery
rails. Everything should work with any modern verson of jquery-rails.

This makes the experience of creating a new solidus app better as the
Gemfile.lock from a fresh rails install won't conflict with our
requirements.

The version requirement was added originally to avoid an issue in
jquery-rails 3.0.1 specifically (double alert confirmation message).