Fix spec failure from DB order issue in postgres
inventory_units.to_a isn't necessarily the same order as
[inventory_units.first, inventory_units.last]

Since there is no explicit order on the association (and no need for
one), it is valid to return them in any order. First and last in rails
add an ORDER BY id ASC if there is not an existing order.