Replace gem rake tasks
Replaces the individual rake tasks for building, installing, and
deploying gems with a simpler system in the top level Rakefile.

Existing tasks were a mix of the PackageTask shipping with rubygems, and
the bundler supplied GemHelpers. PackageTask worked fine, but did very
little, and had a lot of output and temporary files left in pkg.
GemHelpers were installed on all projects other than sample, but
supplied a rake release, which did entirely the wrong thing for our
project (tag and push to github, but only release one of the gems).

The replacement should be simpler, always building/installing/releasing
all 6 gems, and keeping their generated gems in the same place (toplevel
pkg), and should remove confusion about whether we're building the
toplevel solidus gem, or all of the project's gems.