Add hack for specifying user class in association.
Unfortunately, it is possible for classes to get loaded before
`Spree.user_class` has been set in the initializer. As a result, they
end up with `class_name: ""` in their association definitions. For
obvious reasons, that don't work.

For now, Rails does not call `to_s` on the class passed in until
runtime. So this little hack provides a wrapper around
`Spree.user_class` so that we can basically lazy-evaluate it. Yay!
Problem solved *forever*.