Make state creation in admin more reliable
When the selected country was changed, it sent off an ajax call for the
list of states, and replaced the id contained in the new state button.
The new state button sent an ajax call, which responded with javascript
which would hide the button and add a form onto the page.

This had an issue where if you click the add state button, and then
choose a different country, both the new state button and the new state
form would be removed.

The new behaviour is to always have a (possibly hidden) create state
form, and to update the action as necessary. This makes the listing of
states the only asyncronous part of the page.