Fix race condition in frontend address js
Creating the state selector for the current country requires an ajax
request. When the response was returned, the value of the country
selector was used at the time the response was started, even though it
may have changed in the interim.

Updated the js to always use the current value of the country field when
filling in state information.