Prune performance section
I believe this is the one good piece of advice from the performance
section in the README. Asset recompiling has been faster in recent
versions of rails, doing an asset:precompile in development should no
longer be necessary.

I've left in the suggestion to try `assets.debug = false` because I
usually run that locally, and it does make a difference.