Adjust order submenu permissions.
This contains a number of changes to the order submenu to toggle
visibility of certain links. We want users to be able to see the order
but not make changes.

The most notable change is the change around short
shipping. Short ship is the actual permission checked when performing a
short ship, the previous check on cancel was only checked here, and
could be bypassed.