This change disables the ability to change a store credit's category once it's been allocated. If a store credit starts out associated with a certain category and is changed after an accounting reconciliation, the store credit will be incorrectly classified the next time a reconciliation is performed. When a store credit is created with an incorrect category, the admin should invalidate it and create a new one.
When changing a store credit's amount or when invalidating store credit, the admin will have to select a reason why they are performing the action. This helps to better determine a store credit's history.

A new detail page is introduced to display information about a store credit, including it's associated store credit events. This will help admins by enabling them to view what modified a customer's store credit, when that modification was made and why it was made.