Fixes to admin/orders/line_items_spec
These two specs failed very occasionally.

Removing the within_row of the first spec fixes an ObsoleteNode error,
which is caused because the page reloads after changing the quantity.

In the second spec, asserting on #order_total forces capybara to wait
for the page to reload before running the next assertion. Otherwise it
returned true immediately, because the row is removed from the page
immediately after clicking.