Add an Nth order promotion rule.
When this order, will be the nth order the user has ordered, allow this
rule to pass.

It does this by taking a preference for the nth order, e.g. every two
orders. If the user has one complete order, this order would be the
second order, therefore making this rule eligible.

It only takes a users -complete- orders into account. They may or may
not be shipped or paid.