Ensure root redirects work for unauthenticated users.
Previously, if you were not logged in, this would redirect you to the
dashboards url, which would redirect you to the login url.

After login you'd be redirected to where you came from, being the
dashboards url, when this user could have been redirected to the order
url.

By raising here to invoke the unauthorized exception handler, we get
redirected back to RootController#index instead and we can make a proper
decision on the users final destination.

This also prevents the unauthorized flash from showing up after
successful login.