Remove touch from option_value to option_type.
In the case where an option_value is touched, it touches its variants
which touch the product.

In the case where an option_type is touched, it touches the product.

In the current state, when an option_value changes, it ends up touching
the products (and any product touches) multiple times, which is quite
costly, and I'm having trouble thinking of a case where an option_value
needs to touch an option_type directly (since cache is generally at the
product level).