Remove the 'rake' spec shared context
All that's required for spec'ing a rake task is:

    let(:task) do
      Rake::Task['xxx']
    end
    before do
      Rails.application.load_tasks
      task.reenable
    end

That's not much to copy-paste between specs and the readability seems
nicer in this case than the brevity of:

    include_context 'rake'

where we're talking about specs, which we want to be really obvious
and readable.

We're already doing it this explicit way in a couple of rake specs.