Remove Address has_many relationships
With the recent Address work the Address model is meant to be an
immutable data object that isn't viewed as something with (outgoing)
relationships.  The relationships removed here were never used in
Solidus itself and I've never seen them used in an extension or app
and I can't think of any reason why they would be.  Plus, avoiding
defining unused relationships lowers the surface area of the Solidus
api and may help increase security in some scenarios.