Handle edgecase where a user cannot have all roles removed.
When updating a user, I cannot remove the last role from that user. This
is due to an assumption that was made that the form still submitted the
param `spree_role_ids` with an empty array if none of the checkboxes
were checked.

This assumption was false. We have to add behavior around whether or not
the roles can be cleared if that parameter is not present which is not
ideal as explained in the comment in users_controller#set_roles.