Move params massaging code to the controller level
And add better tests.

The only reason we need this code is because our frontend html forms
send data in an awkward way.  That's something we should handle at the
controller level.

This helps simplify some of the model-level code around order update
attributes and should help simplify other upcoming refactoring in that
area.

For some reason we've also been accepting params in the same awkward
way in Api::CheckoutsController#update even though there's not a good
reason to do that as far as I am aware.