Move specific payment amount massaging to the controller
This is another thing that doesn't really make sense at the model
level so move it to the controllers.  We also only apply this in a few
particular controllers (the ones that currently use update_from_params)
so move it to to those particular places where we are using it.