Register extension abilities after permission sets
The order of can/cannot defined by abilities matters. PermissionSets
should never use cannot, because the order they will be evaluated is
unpredictable. We only use them in once place currently, and that should
probably be changed.

Because of this we should evaluate the registered abilities last, so
that any cannots in the custom cancan rules will take precedence (as
they would have before the addition of PermissionSets)