Fix State when config.address_requires_state=true
When config.address_requires_state = false, no dropdown for State
is shown on the Address form. When config.address_requires_state =
true, the dropdown is shown.

Whether the input in the dropdown is required is then controlled
by the setting of @country.states_required = true/false.

However, there was a Javascript check which would make the State
on the form permanently required as soon as any country which
required states was chosen in the Country dropdown.

This change properly marks the field as required or optional as
the Country dropdown changes, and it shows or hides the "required"
symbol appropriately.