Fix use of with_model for MySQL
PostgreSQL can make schema changes, like creating a table, in a
transcation because it is just that awesome. MySQL is somewhat less
awesome, so any schema changes are an implicit commit. This breaks tests
which are run within a transaction.

with_model makes changes to the schema, but if we run it in a
`before(:all)` or `before(:suite)` block, it will be run outside of a
transaction, and it already knows how to cleanup after itself.