Always pass the new order number on unreturned exchange failures
Currently, anything that's not specifically a `ChargeFailure` will not
include the new order number. That's a pity, because something like an
out of stock error shows up as an exception that is not a
`ChargeFailure` and then it's really hard to even find the new error.

This always makes the new order available to the outside so that we
don't have to specifically read it from the exception.