Use the order updater to update the shipment state.
We don't need to do this ourselves.

Additionally a user may define hooks, which are run by the order
updater that wouldn't get triggered whenever shipments are shipped,
which is something worth knowing about.

We also don't necessarily know that updating the shipment state
shouldn't update other states or parts of the order.