Restrict user management permissions.
Only allow users with the UserManagement permission to edit themselves
or users who have no roles.

There is a potential for privelege escalation if we allow users to edit
user accounts that have roles, whereby they change the other user's
email to an email they have access to, thereby gaining access to a
higher-level account.

Given that we don't have a way of knowing the hierarchy of roles, it
seems safer to only allow super admins to edit of user accounts that have roles.