Use helpers from RansackableAttributes on the user model.
At the time this was introduced, the RansackableAttributes module was
not included into the user model. It probably could have been, but
https://github.com/solidusio/solidus_auth_devise/commit/2a5ab50368c9f196abff72b226b302187e19642a
brought my attention back to the notion that we were overriding these
whitelists differently on the user.

This change will make it easier for additions / overrides to the
whitelisted user attributes for people in their specific stores.