Cleanup of shipment spec
Revised specs are largly based on Alex's work, found
https://github.com/solidusio/solidus/pull/49

The shipment spec in the master branch has advanced considerably to
Alex's work with regards to the state-machine's transitioning spec to
`resume`; therefore, the specs from the master-branch have replaced his
work.

Both the state-machine transition specs for `cancel` and `resume` have
been wrapped in a context block, so as to provide a reason as to why these are
a departure from the usual naming scheme. This is because these methods
are not explicity defined, and are defined via the state-machine itself.

* Add spec for #final_price_with_items
* Add spec for `editable_by?`