Apply store credit to orders in the “confirm” state
Currently if an admin allocates store credit to a customer and that customer has already made it to the confirmation step, that store credit will not apply to the order. This change applies the newly allocated store credit to any orders in the confirmation step. It returns the orders to the payment step and re-transitions them to “confirm” using the state machine in order for all of the hooks and callbacks to fire.