Variant image rules
Begin using rules to administer how images are associated to
products/variants.

The current way to associate images to variants forces admins to
upload the same image multiple times if they want to associate an
image with multiple variants.

This change introduces variant image rules (similar to variant property
rules) which allow an admin to associate an image to multiple variants
at once, based on option values.

Taking for example a store that sells t-shirts, with a product that has
multiple variants that vary based on size and color. With variant image
rules, the admin can define a rule that will apply an image to green
shirts of all sizes, rather than upload the same image for the small,
medium and large green t-shirts.

The rules can be as specific or as generic as desired by the admin
(rules without any conditions will apply to all variants).

The change doesn't remove the current way of managing images but it does
introduce deprecation warnings, displays a warning in the current image
management admin page and provides a rake task to migrate current
variant images to the rule based logic.