Rename left sidebar-left to admin-nav
To distinguish the left sidebar (main navigation) from the right
sidebar (3rd level of navigation which appears in the content area)
we will refer to it as the .admin-nav

This also fixes a css spillover from using the class .sidebar which
was causing the right sidebar to be misplaced.

We will, however, avoid renaming templates when at all possible
to prevent breaking users Deface overrides.