Use git ls-files for the gemspec file listing
Currently the new image factory doesn't work in projects because the
image isn't being bundled into the gem.

This ensures we're bundling all the files in the repo, and it makes it
harder to accidentally include files in a release. One change is that
this will include specs in the gem where it didn't before, which IMO is
fine.