Tax Rate refactoring
This looks like a very large commit, but it actually mostly removes lines.

The goal is to make somewhat more understandable what's currently going on in the Solidus
codebase: What rates apply to my order's zone? Those which have shared members with my zone.
When is there a refund? Whenever the store is sending stuff to a place outside the default zone
that does not have included taxes.

The refactoring now implies that a zone have members to be applicable. In the real world, that is
always the case.

There is an expectation for a negative included tax in the reimbursement tax calculator. That
expectation is IMO bogus, I have no idea why it should be there.