Stop using request referer in admin product controller
Prevent setting `session[:return_to]` to a user-inputted value.
It originally tried to ensure that the "Back to Product" link on the
edit page would be preserved with the previous product url query from
the index page, but since the url is saved in `session[:return_to]`
in the index action, setting/protecting from it not being set isn't
neccessary here.