Remove silly broken assignment in spec
This is trying to assign the payment's source to source. Not only is
that the wrong thing to do (a payment is the source of a log_entry), but
since it's being built through the payment.log_entries association, it
will have no effect.

Sometimes two wrongs do make a right!