Make it easier to sort to end of a list
It can be tricky to drag a sortable to the final position of a list.
Because of the placeholder empty lists we have placed there, it
always wants to jump *into* the last item and become a child.

`.ui-sortable-over` is applied to all of the ancestor lists of
the placeholder to give a "lower edge" target (padding-bottom)
and make it easier to sort to the last sibling position.