Simplify inclusion of deleted associations
Previously we overrode association methods like the following:

    belongs_to :variant
    def variant
      Spree::Variant.unscoped { super }
    end

This commit accomplishes the same using a scope on the belongs_to

    belongs_to :variant, -> { with_deleted }

I believe this is shorter and reads better. I think this has only worked
in recent rails version which is why it was not done this way
originally.

This is also significantly faster when accessing associations which are
already loaded. This operation is extremely frequent in parts of the
codebase like Stock.

    Benchmark.ips do |x|
      variant = Variant.new
      x.report("variant.product") { variant.product }
    end

Before:
     variant.product    104.842k (± 2.7%) i/s -    529.537k

After:
     variant.product    737.723k (± 3.3%) i/s -      3.691M