Stock::Packer should not raise InsufficientStock
Stock::Packer currently raises InsufficientStock under specific
circumstances:
* There is a stock item for a variant in the stock location.
* There is exactly 0 inventory remaining for that variant in the stock
  location.
* The variant does not allow backordering.

This is odd, because it doesn't raise InsufficientStock when there is
some but insufficient inventory (it adds the amount available to the
package), or if the stock item doesn't exist at all in that stock
location. These three cases should probably be handled in the same way.

The way the Packer is used is to build the fullest possible package for
this stock location. As it's possible to fulfill an variant half from
one location and half from another, this method should not be raising
InsufficientStock because it has insufficient information to do so.