Add Spree.solidus_version
Spree.version is no longer appropriate as we are not spree and have a
different versioning scheme.

This removes the SOLIDUS_VERSION file and instead adds a
Spree.solidus_version method which all of the gemfiles use to determine
the current version of solidus.

I wish we'd added this method in the 1.0.0 release, but I think this
is better late than never.