Fix payment_with_refund factory
Previously this factory would fail to save because the `refunds`
association was invalid. This was due to a validation failure on the
associated refund because the amount on payment defaulted to nil. This
resulted in the refund_amount (5) to be greater than the payment amount
(nil).

This changes the default the amount of `payment_with_refund` to be
amount specified for the `refund_amount`.