Add image_or_default
This method's goal is to deprecate Spree::Variant#display_image

display_image's goal was to allow the caller not to have to handle the
case where a variant did not have an associated image by creating a
Spree::Image.new and passing that back.

This left too much responcibility in the Spree::Variant class and
eventually led to us taking a boolean param to decide whether or not to
fallback to product image if no variant is available, farther muddeling
the API.

Moving this to a helper puts it closer to the display logic (which is
the only reason we'd want a Spree::Image.new) and more appropriate for
the call.

My bad.