Mark Variant#display_image as deprecated
This was my fault, I though the Spree::Variant creating its own fallback
image if none existed would be a good move, it was not.

It put too much responcibility inside the variant and made it too hard
for callers to know where images were coming from (the product? the
variant? either?) and got more complex when the fallback flag was added.

Rather than using this, which was mostly to support view logic, the new
calls in Spree::ImagesHelper(#spree_image_tag|#image_or_default) should
be used.