Remove Order::CurrencyUpdater
This was intended to be used to update line items in an order to a new
currency. Nowhere do we allow reassigning a currency to an existing
order, or to line items, so this should not be needed.

I don't think we should support changing currencies at all. If this was
desirable behaviour for a store, they should implement it themselves as
there are many caveats.