Mark Spree::Product#display_image as deprecated
Just like Variant#display_image, this was a mistake to add to Product as
deciding the fallback Spree::Image.new should not happen inside of the
Product class.

Callers could not know where this image came from (the master, or a
variant), or could not easily know if no image existed when a blank new
one was returned.