Fix TaxRate.match under ruby 2.3
The behaviour of delete_if has changed in ruby 2.3. Previously, in ruby
2.2, it would delete the items from the Array immediately. Now, in ruby
2.3, the items are removed after iterating over all the items.

This replaces the #delete_if loop from TaxRate.match with more explicit
behaviour with a more explicit #each loop. This works in all versions of
ruby we support and makes the behaviour of that method more clear.

See https://bugs.ruby-lang.org/issues/10714