Always use Zone#contains? for checking tax rate applicability
There's a bit of confusion around when a tax rate is applicable in the code.
Sometimes the order tax zone has to be equal the tax rates zone, and
sometimes it's ok if the tax rates' zone contains the order tax zone.

Using the shortcut from the previous commit, this can be unified.