Allow Zero dollar order adjustments
Sometimes the current value of a promotion is zero dollars, however with
the return check here the Spree::OrderPromotion will not be created and
so the promotion can never be applied in the future by order.update!

This check was added in spree/spree#17c59ec as an aesthetic change which
should not affect promotion logic this deep into the modeling.

A more appropriate place that change would have been a view override of
some kind. As mentioned by @jormon here:

https://github.com/spree/spree/issues/3966#issuecomment-47644133