Fix rake test_app at top level
As of ba2fb4bc108fda51a944455da5c60807845ef007 we check if the various
engines are defined when building the initializer. This means that
running rake test_app at the top level now builds the initializers
incorrectly as all of those are defined at the top level.

This fixes that issue by re-running rake test_app in each subdirectory.