Fix rails reloading issue for StockConfiguration
For some reason, rails becomes confused when this is required in
AppConfiguration and remembers the Spree::StockConfiguration module as
an autoloaded constant. This means that it will unload it when code
reloading occurs.

This is fixed by requiring it explicitly in core.rb

Steps to reproduce:

In a fresh sandbox run:

rails c
> Spree::Config.stock # works
> reload!
> Spree::Config.stock # errors

Fixes #725