Make StockConfiguration a namespaced class
Replaces the Spree::StockConfiguration constant with a
Spree::Core::StockConfiguration class.

We should use classes whenever possible rather than modifying constants
and module variables. This better matches AppConfiguration being a class
(Spree.config being the instance of that class. This also makes the test
cleaner.

The name is changed to match the file it is in.