Use Spree::TaxRate.for_zone scope in Spree::TaxRate.match
This is really an intermediary step. I'm using for_zone already in
Spree::TaxRate.match, which should result in a constant eight DB calls per
.match call. Previously, the number of DB calls would depend on how many
tax rates there are, and how these are configured (as Spree::Zone.contains?
would be run for every tax rate in the system).

A side effect of this refactoring is that the reimbursement specs had to be
amended a little: The zones had to be created `:with_country`, because otherwise
there'd be no shared members and the tax rates wouldn't apply. This actually makes
the tests slightly more realistic (what is a zone without members, anyway?).

Also, I had to change an expectation: The included tax total of the reimbursement
was previously expected to be negative (which is something that IMO should never
happen, as negative included taxes at the moment become positive additional taxes).