Handle default VATs and tax rates for order separately
Before this commit, the taxation system would handle default VATs and
rates that *actually* apply to the order's item with the same code.
However, they do fundamentally different things: For example, the
`pre_tax_amount` should only be calculated from default VAT rates.

This commit introduces a new private method on the ItemAdjuster to
handle "deductable" VAT rates. This fixes some of the taxation specs,
because now the pre tax amount is calculated correctly, and it opens the
path for a cleverer handling of default VATs (namely, changing the line
item's price).

Additionally, this commit introduces cached options for the entire order
to be passed into the item adjuster.