Add additional guards to Spree::Shipment#tracking_url
A tracking number, and a shipping method are both prerequisites to
generated a proper tracking url. I chose to return nil here, if the
tracking number or shipping method is not yet set.

I did this because it mimics the behavior of
Spree::ShippingMethod#build_tracking_url when tracking is blank.