Add taxation-relevant columns to Spree::ShippingRate
Adds `included_tax_total`, `additional_tax_total` and `pre_tax_amount` to
Spree::ShippingRates.

`promo_total`, which would have also been possible, is implemented as a dummy setter
and a getter that always returns zero. Shipping rates do not have promotions applied
to them, that is what happens when they are used for a shipment.