Refactor order factory
Add `with_payment` and `line_items_can_ship` traits

Adds two traits to the base order factory. These will make it easier to
create orders that can actually ship (that haven't already been
completed).

These traits are based on the order_ready_to_ship factory with the
difference being that the payments are lazy evaluated so that they can
be easily overridden in a test instead of just existing in an after hook.

payments that were manually created in this file no use the trait.

e.g.
====
     create :order_with_line_items,
            :with_payment,
            payment_is: 'completed',
            payment_type :credit_card_payment

     # or just use defaults (completed, credit card)
     create :order_with_line_items, :with_payment