Split the skeleton CSS into separate files
We need to make changes to how the skeleton operates in the backend with
the addition of the left navigation. Unfortunately, the changes required
would apply to both the admin and the frontend so we need to separate
these files.

Ideally, the admin and the frontend wouldn't share any assets at all
which this commit also pushes us in the direction of.