admin/search#users should be top-level json object
Returning a top level array is vulnerable to JSON hijacking, as the
JSON can be interpreted as valid javascript.

http://haacked.com/archive/2009/06/25/json-hijacking.aspx/

Modern browsers do not allow intercepting array initialization, and also
should not allow execution of responses with json content type. Rails
adds X-Content-Type-Options:nosniff by default, which protects browsers
as old as IE9 from the vulnerability.

If this was exploitable, a malicious website which was visited by an
admin could reveal users's email, shipping, and billing addresses.