Add ability to filter ids api resource controller
A number of existing controllers support this feature: taxons, products,
option_values, option types, properties. Adding this to the resource
controller allows those controllers to be converted to a resource
controller, and adds consistency across various endpoints.

This could have been done previously with ransack, using attribute_in,
so there is no change in information we expose, just allowing another
means to request it.

This is also helpful when using the API through select2 to populate
the data from the already selected ids.