Match rates in the order and line item adjuster using address
This is the first step in removing the notion of the default tax rate
and the *very* complex code required to find it.
I'm using the new TaxRate.for_address scope to find tax rates for the
default address (Spree::Address.build_default, it's does not have to be
persisted as I really only need the country ID, which is present).

I had to adjust some specs, in which the default_tax flag on zones would be
abused to just apply to anything (which, I think is the source of some of the
confusion regarding that flag).

I deleted the `precision of pre_tax_amount` specs on shipment and line item.
These specs will very soon be unnecessary as the column will go away as
```
amount - included_tax_total == pre_tax_amont
```