Match tax rates using address instead of using zone
I'm using the new TaxRate.for_address scope to find tax rates for the
configured default address.

I had to adjust a lot of specs, in which the default_tax flag on zones would be
abused to just apply to anything (which, I think is the source of some of the
confusion regarding that flag). Some spec setups have now gotten a bit longer,
but there's an upside: They've also become more understandable.

For the time being, I've left the `default_tax` boolean in the factories and have
not endeavoured to remove all the code that possibly could be.

Return the configured default tax address if the order's tax address is nil

Instead of first running `Spree::Zone.match`, which is an expensive method to call,
we can just return the configured default address (which might well be `nil` for
our American friends) as the fallback on `Spree::Order.tax_address`.