Use TaxRate.for_address in shipping rates
The shipping rates were still simply checking whether their rate's zone is
equal to the tax zone, and not whether maybe the address in question belonged
to a zone that *would* have an included tax rate.

Using the .for_address scope, I could fix these, so now also for shipments where
the rate zone is europe even though the smallest matching zone for the address
would be Germany, this works.

Also, I made the pre_tax_amount non-rounded (as that produces rounding errors
lateron).

To make sure the taxes for a cart without addresses are done correctly, the commit
adds a fallback to Spree::Order#tax_address: The default tax address.