Make shipping rate adjustments non-eligible
There is a concern over shipping rate adjustments messing up
shop reporting - as those do not actually count towards the order total,
but are actually more estimations of what's going happen if people
use the `Spree::Order#all_adjustments` association, which the shipping rate
adjustments would be part of.

We had earlier discussed setting the `order_id` for those adjustments to `nil`,
but that makes the shipping rate adjustments invalid. Setting the `order_id` on
adjustments to `nil` would require the presence validation for the order to be
conditional on something - in effect, touching the `Spree::Adjustment` class,
which is not a beast I want to touch here.

Now, there is a documented Boolean on adjustments that literally says:

>  1. *eligible?*
>
>     This boolean attributes stores whether this adjustment is currently
>     eligible for its order. Only eligible adjustments count towards the
>     order's adjustment total. This allows an adjustment to be preserved if
>     it becomes ineligible so it might be reinstated.

While this was initially conceived for promotion adjustments, it can just as well
be used for tax adjustments.

This commit adds new instance methods on taxable items (Spree::LineItem, Spree::Shipment and newly
Spree::ShippingRate) indicating whether they contribute to the order total. If they do, their
adjustments surely must as well, and if they don't (as in the case of shipping rates), they
answer `false` when asked: Are you `eligible?`.

While this is not a cure-for-all (If your reporting does not honour the `eligible` bit, you'll
still have wrong numbers in your tax reports), I think it provides the necessary
distinguiability for reporting.

Also, the complexity increase for the taxation system is small, and the adjustment system in
general is not even touched.

The order updater explicitly only calls adjustments on line items and shipments, or adjustments that
have the order instance has adjustable when calculating the `adjustment_total`.
`Spree::Order#all_adjustments` is only executed for re-calculating adjustment amounts.