Add Spree::Config.default_vat_country_iso
Use the new Spree::TaxLocation class with the country defined by
Spree::Config.default_vat_country_iso to specify the area for which
taxes are calculated when an order is in `cart` state.

This relieves us from the following issues:
- We don't have to assume the database to be present on app boot time
- We can tighten the contract for the tax adjusters, because now
we know that the tax address will always be there, but could be
a Spree::TaxLocation.