Allow option_values_ids to be an array
Previously the permitted variants on an attribute would only allow for a
scalar value.

This changes it to allow an array, but will have the side-effect of no
longer allowing a scalar to be set. I'd suggest this is a  pretty safe as
nobody should be sending _ids a single value, while we're fixing
something that is a bug.