Remove ShippingMethod.calculators override
This was overriding the default implementation in
Spree::CalculatedAdjustments, to select only calculators which were a
subset of Spree::ShippingCalculator.

There is no reason to do this, as calculators are configured on
Rails.application.config.spree.calculators.shipping_methods

Having this caused shipping calculators to disappear if code reloading
caused ShippingCalculator to be reloaded.