Do not round pre-tax amounts
When rounding the pre tax amount for VAT calculations,
we run into rounding errors on creating the adjustment.

Example: We want the tax amount for $16.00 at 19% VAT
16.00 / 1.19 = 13.4453 ~ 13.45 --> 2.55 VAT
13.45 * 1.19 = 16.0055 ~ 16.01 --> 2.56 VAT (wrong wrong wrong)

Not rounding the pre tax amount fixes this rounding error.

As usual with rounding, there will still be cases where this does
not work, depending on how many digits our database stores. They
will be much rarer, though.

There is a failing test for this, but it's pending for a different
reason (in the taxation integration spec, we have the aforementioned
rounding error, but it doesn't show up because shipping rates are not
tax adjusted because of a missing `contains?` check.