Change product variants index page logic
Currently there is a message and link to create option types even if
some exist but are just not linked.  This will leave that link if no
option types actually exist, give a link to create a variant if any
option values are defined on the product, and if none are it will
explain that while giving a more obvious link back to the product
details where option values can be added.