Fix deleting tax adjustments in Spree::TaxRate.adjust
As reported in https://github.com/solidusio/solidus/issues/909,
when calling Spree::TaxRate with a mixture of shipments and line
items, it will fail to delete some tax adjustments (those that don't happen
to be the type of the first one).

While fixing that, I realised that this code would set the pre tax amount for
untaxed items to zero, violating that the pre tax amount for something
that does not have included taxes should be equal to its discounted amount.