set billing address to default address
Set `bill_address` to default address, if user has one.

Currently when order goes to `address` state, only ship address gets filled in from a previously entered address. This is because `ship_address` is set to default address because `ship_address` is an alias for `default_address` in the `UserAddressBook` module.
https://github.com/solidusio/solidus/blob/master/core/app/models/concerns/spree/user_address_book.rb#L53
However bill_address is not an alias, so it is not prefilled and user has to enter it every time. Rather than making it into an alias, just try default_address in `assign_default_address` if user has one.