Utilize the checkout tabs partial
This will replace the configurations right nav menu on the adjustment
reasons, refund reasons, reimbursement types, and named types page with
a nicer tabbed panel.  Currently the named types partials are only
utilized by return reasons.  It appeared to have been an idea to DRY the
code but the logic breaks down in places and it might be worth it to
remove it going forward as this introduction of the tabbed panels is
another place where its generality breaks down.