Add vat_country relation to prices
I chose to use the ISO code as primary/foreign key as that is just so
much easier to understand than the ever-changing country IDs.

I believe with the 2-character limit on the column and Indexes that
should not be a major performance issue, and it makes our database dumps
and SQL queries a lot easier to read.

I don't want to touch the default price and would like that to always be
the net (export) price, as that makes prices easy to compare.