Add valid_from column to Spree::Price
The idea here is to introduce a valid_from column to prices that
determines which price is taken if there is more prices that fit the
given search scope.

For example, if you have two prices for currency JPY:
```
a = Spree::Price(currency: "JPY", amount: 13, valid_from: 2015-12-31)
b = Spree::Price(currency: "JPY", amount: 13, valid_from: 2014-12-31)
c = Spree::Price(currency: "JPY", amount: 13, valid_from: 2016-12-31)
```

And you want to find the price currently valid on Feb 2nd, 2016, you would
obtain price b - the latest price that is not valid in the future.

If you wanted to find out which price *was* valid on Feb 2nd, 2015, that would
be price a.

You can even find out a valid price far in the future, although that's more
brittle, as is the case with most telling the future things.