Ensure tooltip dissapears when canceling splitting
Since switching to bootstrap tooltips, canceling a split shipment left a
hanging cancel tooltip. This manually closes the tooltip when canceling
a shipment. I didn't see anywhere else in the codebase that this was an
issue.

Bootstrap proposes to solve this before the 4.0 release using
MutationObserver (see https://github.com/twbs/bootstrap/pull/17021).

This was not an issue with powertip, which uses a setInterval to check
if the tooltip is still valid and being hovered.