Add LegacyLineItemPricer
This moves setting the price of a line item to a special service object
called a `Pricer`. It supports price modifier options as well as not
changing the line item if no price modifiers are set and a price and
currency are already present.

Previously, the currency of a line item could be obtained from the
variant (which really can't know about the currency of a line item), so
that is newly streamlined into obtaining the currency from the if the
options don't say otherwise.