Ensure requests finished at end of spec
Without this it is possible for a request to be served at the same time
we are using DatabaseCleaner to truncate the tables.

In a future version of capybara this won't be necessary as it will on
pending connections, but we might as well do this as it asserts that the
operation actually worked.