Get line item currency from order
There seems to be a way of obtaining a line item's currency from
it's variant. That doesn't make much sense, as a variant might have
many prices in different currencies, but is not tied to a user or an
order which would know WHICH of those currencies to use.

Yanked an occurrence of `variant.currency` for `order.currency`.