Introduce PriceModifierLineItemPricer
There is an extension point for modifying the price of a line item
that is used in the following extensions:

* https://github.com/godaddy/spree_product_sale
* https://github.com/v-may/spree_simple_sales

This extension point introduces some rather interesting code to
Spree::LineItem, which this commit extracts to a pricer object.

The extension point, as far as I can overview, needs the following to work:

- `*_price_modifier_amount_in` methods defined on the variant
- Fitting `*` methods on `Spree::LineItem`
- The Line Item to accept those methods as params

This commit extracts that code from Spree::LineItem, with the intention
of providing an example for how to write a custom pricer as well as with
the intention to remove this pricer and the associated extension point by
Solidus 2.0.