Remove LineItem#currency column
Somehow, in the past, it was possible to have orders with line items
in multiple currencies. The code to remedy this (Order#homogenize_line_item_currencies)
disappeared in #635, and now we have a bunch of code in different places
that allows setting it, raises errors if it's not there, and so on.

This commit delegates the line items currency to the order.
Whenever people try *setting* a line item's currency, they get a
pretty deprecation warning telling them not to do so.

The error handling functionality that seems to be used in the order
importer is still present (ie: If we know an imported line item has
a currency different from the order, we'll error out).

This commit replaces all that behaviour with a simple delegate that
always obtains a line item's currency from it's order.

Adds a changelog entry.

I also refactored `OrderContents#add_to_line_item`, taking code from
https://github.com/MountainRoseHerbs/spree/commit/461f97510c29d50c08c6d8b75805b66f12de12af

That method should now be much easier to understand.