Always build a new price object when using Variant#price=
We want prices to be essentially immutable, and we want a history
of prices for our variant. This commits makes it such that the present
default price is replaced instead of modified.

Unfortunately, the old price is also deleted in the process. ActiveRecord
does that, and there does not seem to be any way of stopping it!

Fixed spec setup for product helper spec.

Do not set currency in product duplicator (the master price will take care of that itself)
Do not set variants default price currency, as the price object will handle that.

Fix another spec so that the validation for price (which isn't tested) does not barf.

Use the correct scope instead of unscoping `default_price`

`acts_as_paranoid` offers a `with_deleted` scope for paranoid models.
This commit uses that scope in the `scope` argument of the `has_one`
declaration instead of using `unscoped` and overriding the generated method.