Encapsulate the logic for when an item is shippable.
Transitioning an item (InventoryUnit) to `shipped` is restricted
allowing an item to ask itself if it is shippable before it is
marked as shipped[1]. This questioning is implemented by verifying
the item is on-hand[2].

But when the overall shipment is checking to see if the item is shippable
and it does it's own interrogation of the items in the shipment rather
than asking the items itself[3].

Worse is that the implementation differs from the implementation in
InventoryUnit. InventoryUnit questions if the item is "on-hand" while
in Shipment it questions if the item is not backordered. Unsure if it
is a real issue but it might mean we could get a case where the inventory
unit is in neither state allowing the shipment to think it can ship
while the inventory unit thinks it can't resulting in an error.

This commit encapulates the decision entirely in InventoryUnit allowing
the item to determine if it is ready to ship. This logically should
be the same as before only we now have better encapsulation.

The additional benefit (and primary motivation) is that the condition
can now be customized now that it is properly encapsulated. I am working
on an app with on-demand manufactoring with no real inventory. An item
cannot be shipped until it is built. We can customize InventoryUnit as
follows:

    Spree::InventoryUnit.class_eval do
      module CheckBuiltForShipping do
        def allow_ship?
          super && built?
        end
      end
      prepend CheckBuiltForShipping
    end

Prior to this commit, the system would allow the user to attempt to ship the
item (since it is not "backordered") but the ship would fail since our
customization made it not shippable. With this commit the shipment it now not
shippable also meaning the user isn't even presented with the option.

This commit also affects the `determine_state` method to use the same conditional
methods that are in the state machine to further the goal of having a single
source of truth. This is necessary as the order updater bypasses the state machine[4]
and uses `determine_state` instead. One of the conditionals was also modified so
that the `allow_ship?` is used again for a single source of truth although an exception
is put in place so that a cancelled unit doesn't prevent moving to the ready state.

In theory this commit should have required not test changes as it is just
refactoring implementation details. Unfortuantly due to the abuse of mocks
there are some adjustments to the test to get things passing.

1. https://github.com/solidusio/solidus/blob/master/core/app/models/spree/inventory_unit.rb#L58
2. https://github.com/solidusio/solidus/blob/master/core/app/models/spree/inventory_unit.rb#L118-L120
3. https://github.com/solidusio/solidus/blob/master/core/app/models/spree/shipment.rb#L84
4. https://github.com/eric1234/solidus/blob/truegault/core/app/models/spree/shipment.rb#L320-L324