Re-add a dummy currency setter on variant
It does not make any sense to set a currency on a variant - if it's the default
currency, it does not change what prices do by themselves, and if it's not,
you're changing the default price to something that can not be the default as
it would have the wrong currency.

However, there was a number of places in the old code that still tried to do that,
so I expect custom code to also have that code. This commit re-introduces the setter
with a Deprecation warning.