Bump to more recent version.
Solidus was bound to a version several years old. I'm actually a bigger
fan of just letting it float (especially since it is a dev tool) but
I know some people think all dependencies should be bound to a major
version and that seems to be the pattern in Solidus so following that.