Use with_deleted as scope for Variant#prices relation
We want to know all prices for a variant. When building a new `default_price`
as we do every time `Spree::Variant#price=` is called,
ActiveRecord will silently `destroy` the old price. `Acts_as_paranoid`
mitigates this behaviour, and the `with_deleted` scope here appears like
it never happened.

I briefly considered overwriting `Spree::Price#destroy` to be a noop,
but that would be an introduction of where uncommon behaviour for a model.