Make the creation of the sandbox consistent with the dummy app.
The dummy app allows a DB_HOST variable[1] to be set when creating
test apps and the database username is `postgress`. The sandbox app
on the other hand only allows the database type to be set. It is
always assumed to be `localhost` and the user is assumed to be `sandbox`.

This commit makes the sandbox app consistent by allowing the DB_HOST
variable to be set and it uses the same postgres user so that the a
new special user does not need to be setup (i.e. we can use the same
user as we are using for our tests).

1. https://github.com/solidusio/solidus/blob/master/core/lib/generators/spree/dummy/templates/rails/database.yml#L30