Add a new mixin to create the line-through effect
This is an effect that we're using in several parts of the admin and
we're using a few different tactics (most with confusingly dirty CSS).
This mixin attempts to remove that duplication and clean up the code.

Moving forward this should be used to create the effect. The effect in
question:

------ Text Here ------