Re-organize admin variant form
Before, the form for editing or creating a variant was organized in two
columns without any grouping or sectioning. This commit re-organizes the
fields into a "general", "option values", "dimensions" and "pricing" section
using `fieldset` and `legend` elements.

Pricing is added at the bottom of the form, so that additional prices can easily
be added lateron.