Only allow editing option values for new variants
Currently, when an admin edits a variant, he or she does not have
direct visual feedback as to what variant he or she is editing. In fact,
she is able to change the variant's options to those of a variant that already
exists.

This commit changes the form such that option values, which define what a variant is,
can only be edited when the variant is a new record. When it's already persisted,
the legend for the form will reflect the variant's `options_text` so that the user
knows what she's editing.