Replace variant index content area with "new" form
Previously, when adding a new variant, the form would be shown
between the table filter and the table, which gave the user all kinds
of confusing options:

- Click search and lose the data they might have input for the variant form
- Enter option values in the search form, thinking they would apply to the new variant
- Click edit or trash in the table, again losing data they might have input already.

This commit reduces the visible complexity when creating a new variant. For the JS to be
simple, I added an ID to the area that holds the content, but no navigation.

Additionally, this commit now hides the table filter search options if there are no variants
(deleted or not) to filter.