Show or hide deleted variants using table filter
Before, in order to show deleted variants, the admin user would press
an "action" button outside the table filter (with a filter icon).

This commit changes that so that a check-box in within the table filter
governs whether to show deleted or normal variants.

Also refactored an instance variable (`@deleted`) away, as it was not
so useful.