Use tabs on product edit page
We have a great new tabs feature to replace the third-level sidebar
navigation. This commit uses the tabs feature on the product edit page.

For the time being, the only point where this messed up the existing layout
was `products#edit`. I solved it by making the first container full-width and
using the new sidebar space for pricing, variant and dimension options.