Use Spree::DisplayMoney to convert Line Item amounts to Money
We have a library to convert BigDecimals to Spree::Money objects for
display purposes. This commit uses `Spree::DisplayMoney` to generate those
methods instead of the line item object rolling its own thing.

This led to a few naming changes, for which aliases are provided.
The main method that's aliased in each case is now the one generated by `Spree::DisplayMoney`.