Simplify Spree::Variant#price_in
The commit changes behaviour. Before, `Spree::Variant#price_in` would
return a price of zero if no currency was given or no price was found in
the given currency. I think that is dangerous: You don't want your products
to randomly be sold for nothing.

This commit changes this behaviour to return the default currency if
no currency is given. It furthermore uses the database's finding capacities
rather than iterating over a potentially large array of prices, speeding things
up.

If no price for the given currency is found, returns nil.

Contains a minor spec setup fix in which a price is properly created.