Backend variant index view: Show deleted variants along non-deleted
Before, when you clicked the `Show deleted variants` checkbox, you would
only be shown deleted variants. The behaviour people would expect here
is actually being shown deleted as well as non-deleted variants, with a
visual indication (red!) which ones are deleted and which ones are normal.

This commit does that.