Allow other admins to see the taxonomies
The taxonomy index already has precautions against letting users edit
and delete taxonomies. There are cases where it will be valuable for
admins without permission to edit or delete taxonomies to see what
taxonomies exist.

This may be a change we should be making in other areas of the admin as
well.