Deprecate currency setter and getter on Spree::Variant
A Variant does not have one currency. A price has that. This commit
deprecates setting and getting a variant's currency.

The old specified behaviour is even a little strange when you imagine
it somehow being called in a context where `current_store` is not the
default store and has a different currency. In that case, `Spree::Variant#currency`
would return the wrong currency (the installation's default one).