secure compare the spree_api_key
When comparing the API keys, Solidus is susceptible to timing attacks. Given a list of ~250 possible tokens:

0000000000000000000000000000000000000000
0100000000000000000000000000000000000000
0200000000000000000000000000000000000000
... snip 246 ...
FD00000000000000000000000000000000000000
FE00000000000000000000000000000000000000
FF00000000000000000000000000000000000000

A hacker can recognize that there was a 100µs difference in string comparison between, say, A1 and F3. If A1 took longer, then we know that there’s a user with a token that begins with A1… Rinse and repeat and you’ll be able to find a token.

Although the internet has a significant amount of jitter, a 2009 paper from Rice University [Crosby, Wallach, and Riedi, 2009] noted that it is possible to achieve 15-100µs accuracy. If the request was over a local network, the accuracy improved to 100ns.

http://www.cs.rice.edu/~dwallach/pub/crosby-timing2009.pdf

To prevent this, we use a constant-time comparison algorithm to compare the bytes.

References:

Crosby, S. A., Wallach, D. S., and Riedi, R. H. 2009. Opportunities and limits of remote timing
attacks. ACM Trans. Inf. Syst. Secur. 12, 3, Article 17 (January 2009), 29 pages.