Add Variant Pricer and related Pricing Options class
The varaint pricer class can either be instantiated using a Hash or using an order.
When setting the price for a line item and when displaying the price
of a variant, so far we're assuming that only the currency is relevant
for the price, when really it could be a number of other things.

Calculating a line item price is substantially different from showing
the price for a variant in a frontend helper, as in one case we use
the `current_currency` variable, while in the other we take things from
the order. This class provides a single source of truth for all things
necessary to choose a price in both cases.

To indicate the close relationship of the pricer class with its
options class, the pricing options class is namespaced under the
pricer class.

Added tests to reflect that the option keys have to reflect columns
on the Spree::Price model, as it is going to be used for ActiveRecord queries.

Also added a #cache_key method to replace the use of `current_currency`
in frontend cache keys.