Fix case where `track_inventory_levels` global flag not being honored
At the instance level[1] the system looks both at the variant flag as
well as the global flag, but when operating at the collection level[2]
only the variant flag was taken into account. This commit fixes this
by making the `in_stock` scope a no-op if Solidus is configured to not
track inventory levels globally.

I hit this issue when trying to create orders in the backend while
configured to not track inventory. The product selector makes a call
to the API and indicates the `in_stock_only` flag[3]. Since none of the
products had inventory (since inventory is not tracked) no products
where returned.

1. https://github.com/solidusio/solidus/blob/master/core/app/models/spree/variant.rb#L307-L309
2. https://github.com/solidusio/solidus/blob/master/core/app/models/spree/variant.rb#L68-L74
3. https://github.com/solidusio/solidus/blob/master/backend/app/assets/javascripts/spree/backend/orders/edit.js#L4