Upgrade path for shipping rate tax data
Previously, we had a migration that deleted the `tax_rate_id` column
from the `spree_shipping_rates` table. Within that migration, we could not
satisfactorily create shipping rate taxes because that required a call to
`Spree::TaxRate#compute_amount`, for which we can not guarantee the interface
for all future Solidus versions to come.

This compromise works as follows: We keep the column and provide a rake task to
migrate that data up. In a future release (maybe 2.0), we can then add a migration
to delete the column that errors out if there's still data on it (indicating the
migration hasn't been run).