Implement Pricer class
The pricer class knows its related pricing options class, and gets
default options from there.

Here it also becomes apparent how the pricer's options hash has
to always coincide with column names on the Spree::Price model.

I've kept the behaviour of `price_for` similar to `price_in`, except for
two things: Instead of returning a `Spree::Price` object, return a `Spree::Money`
object (because that's all we need), and of course accepting more options
when searching.

The `is_default` flag on `Spree::Price` indicates whether a price is currently valid,
which is not exactly something that I see in the pricing options (as it's not something
that we would directly search for. I replaced that with a scope on Spree::Price.