Change current_currency to current_price_options where appropriate
This is mostly in caching situations where we want to make sure that
the cache also gets invalidated once we show the page with, say, a
different country.

Adapted some cache keys

Changed the pricers behaviour so it can actually return `nil`, as Solidus
seems to expect that of a pricer.