Add proper scopes to property sorting
Ordered properties need to be scoped to a product (or
variant_property_rule). Otherwise, when creating or updating a property record,
the position field gets overwritten for all properties in that table.

Because of the differing fields in product_property and
variant_product_rule_value, it was necessary to pull the acts_as_list
declaration out of the concern and put it directly into the model.