Do not ajax in the new variant form
While this did save us a page load it also required a slightly hacky
workaround with the `js-content-below-tabs` class. Making this form
appear on a page load is necessary for switching over the index page to
the new layout style where we no longer have `.js-content-below-tabs`.

While we could add that class to the new layout I believe it's better
for us to to working towards a cleaner, more extendable layout.

Additionally, this commit updates some of the hash syntaxes.