Add Prices tab to Products overview
We have great functionality for dealing with multiple prices for each variant. This is an
absolutely basic interface so that as a store admin, it's possible to edit and update prices.

I'm unsure about allowing deletion and have left it out for now: When you delete a price,
you might end up with no price that has `is_default` set, which will then make the variant
in question become "unavailable". As a new price gets the `is_default` boolean anyways, the old
one becomes so meaningless it's as if it was deleted.

There are plans for a much better interface, but I need something simple that works for 1.3
to get out the door.