Rename Promotion#expired? to Promotion#inactive?
The previous method name was not indicitive of what this method checked,
as most people would not consider as-yet-unstarted promotions to be
"expired".

This commit changes the method name to "inactive?", adding a new
implementation for "#expired?" that aliases to the new method, adding a
deprecation warning.

A further commit will address existing usages of the "#expired?" method
from within Solidus itself.