Improve mailer previews DB sampling
Previously these mailer previews just found the first object in the
database and used that to render the preview.

Now they make an effort to find a suitable record in the DB
(completed/cancelled order, etc.) and raise an understandable error if
there is none to be found.

This also now finds the _last_ matching item in the database, which
should allow a developer to, for example, cancel an order and then work
on what the cancellation email would look like.