Remove is_default Boolean from prices
This boolean is either a very bare-bones version of `acts_as_paranoid` or
a very bare-bones version of `papertrail`. However: It's really badly named,
as `is_default` is not something I understand very well at all.

We can accomplish the same behavior by simply using an `order` clause in
the `currently_valid` scope on prices.