Rewrite payment method selection in new payment
Previously a credit card payment method would have to be named "Credit
Card" in order to allow creating new payments through the admin. This is
because it was using the payment method's name, running it through I18n,
and using that as a selector to find a fieldset to show.

Now it just finds a div with the correct id.