Ensure that one store is set as the default
A previous change to the CreateStoreFromPreferences migration caused the
store to no longer be set as a default.

This updates that migration to set the store it is creating as the
default. This also adds a later migration to set the first store to the
default if there is not an existing default. Since the
CreateStoreFromPreferences migration will be run very shortly after the
creation of the Store table, this should be a safe assumption.