Invoke on payment method not payment
Yes, we want to void the store_credit, but the payment method
has a set of complicated action handlers to wrap the store_credit's
response in an ActiveMerchant::Billing::Response

Both #credit and #void were being called on the wrong class, but #credit
actually has a different signature, so I will change it separately. Also
this makes it easier to tell which spec corresponds to which code path.