Reduce stubbing to users_controller_spec
Previously this was stubbing all sorts of ActiveRecord methods to try
and intercept the calls that we thought would do what was intended.

Instead, this spec now allows the controller to operate on real models
and we assert on the result.

The remaining stubbing is on authorization, because that allows us to
explicitly test for the permissions we expect to be checked.