Properly namespace route in navigation helper
This helper would throw an exception if used outside of the Spree
namespace. When adding customizations to the admin you may want to
namespace new sections under your own app to help differentiate
customizations and native Solidus. This change will allow you to do that
and continue using helpers like this.