Make checkout spec from address more realistic
This spec was no longer testing anything because the stub of
Spree::TaxRate#match was no longer being hit. The test was succeeding
due to a bug related to #changed? in ItemAdjustments.

This commit removes some setup which helped the test pass without
actually recalculating tax.

This also removes some stubs in favour of testing the real system.