Convert AC::Params to hash in PaymentCreate
ActionController::Parameters#to_h returns a hash of permitted
attributes. If nothing has been permitted it returns {}.

Previously, we relied on ActiveRecord for filtering, which did work, but
it's simpler and easier to reason about if we filter the attributes
immediately upon receiving them.

This causes a small behaviour change. Previously, if unpermitted params
were passed in it would raise an ActiveModel::ForbiddenAttributesError
exception. Now it will be the equivalent of having passed in an empty
hash.