Cancelling Inventory updates tax_total & item total
Currently, when inventory is cancelled on an order, the system creates a
negative line item adjustment for the the line item total including
taxes. This works for refunding the customer appropriately because the
order.total takes the adjustments into account. However the
order.item_total and order.tax_total still reflect the number of line
items the order was completed with.

I've provided two tests to the order cancellation spec which will pass
when the orders totals are correctly updated.

The question for the community is: Should cancelling and inventory unit
update the order's tax_total and/or item_total and how this should be
implemented with regards to the existing cancellation line item
adjustment.