A hook to pass options to OrderContents#add
In the context of writting a solidus subscription gem which passes along
some extra fields from the add_to_cart form I found that I wanted to
pass extra params to OrderContents#add.

However in order to do this I needed to override the entirety of the
OrdersController#populate method.

This provides a private method which can be overriden with the options
to be passed into order contents.

In my case this would allow me to process subscription options in
OrderContents without needing to completely obfuscate the functionality
provided by the front_end controller

Im hoping this might be of use to anyone else wishing to hook into which
options get passed to OrderContents#add.