Improve testing of payment_parameters
The previous tests for the parameters being unchanged weren't testing
anything since the expected object may have been mutated by the system
under test.

Adds additional tests to ensure that nothing is being permitted when
calling these methods.

This changes the tests to convert the result into a hash rather than
comparing the ActionController::Parameters. This avoids deprecated
behaviour in rails 5 when comparing an AC::Param object with a hash.