UserMethods#on_sign_in method, extracted from solidus_auth_devise
This is not devise-specific code, if it's in solidus UserMethods, then
any auth integration can simply call it, instead of copying and pasting
from solidus_auth_devise.

By calling the method `on_sign_in` instead of `associate_guest_orders`,
the idea is that if solidus changes it's mind about what needs to happen
on sign in, or how it should be implemented -- it can just change this method,
which auth integrations will continue to call with no changes in their code,
no need for new logic to be copied to every existing auth integration.

If this is in Solidus, the [solidus_auth_devise implementation of
the post-sign-in hook](https://github.com/solidusio/solidus_auth_devise/blob/a6e27e98d80f507a7064fdd63f2950f622b7d93e/config/initializers/warden.rb)
could be changed to something like:

    user.on_sign_in(guest_token: auth.cookies.signed[:guest_token]) if user.is_a?(Spree::User)